+++ src/main/java/dao/NoteRepository.java
@@ -1,10 +1,9 @@
 package dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import entity.Note;
 
-public interface NoteRepository extends JpaRepository<Note,Integer>{
+public interface NoteRepository extends JpaRepository<Note,Long>{
 
 }
